+++ cmd/opentelemetry-collector/app/defaults/default_config_test.go
@@ -32,7 +32,7 @@ import (
 	"github.com/jaegertracing/jaeger/cmd/agent/app/reporter/grpc"
 	"github.com/jaegertracing/jaeger/cmd/opentelemetry-collector/app/exporter/cassandra"
 	"github.com/jaegertracing/jaeger/cmd/opentelemetry-collector/app/exporter/elasticsearch"
-	grpcExp "github.com/jaegertracing/jaeger/cmd/opentelemetry-collector/app/exporter/grpcplugin"
+	"github.com/jaegertracing/jaeger/cmd/opentelemetry-collector/app/exporter/grpcplugin"
 	"github.com/jaegertracing/jaeger/cmd/opentelemetry-collector/app/exporter/kafka"
 	kafkaRec "github.com/jaegertracing/jaeger/cmd/opentelemetry-collector/app/receiver/kafka"
 	jConfig "github.com/jaegertracing/jaeger/pkg/config"
@@ -88,12 +88,12 @@ func TestDefaultCollectorConfig(t *testing.T) {
 		{
 			storageType:    "cassandra,elasticsearch,grpc-plugin",
 			zipkinHostPort: disabledHostPort,
-			exporterTypes:  []string{cassandra.TypeStr, elasticsearch.TypeStr, grpcExp.TypeStr},
+			exporterTypes:  []string{cassandra.TypeStr, elasticsearch.TypeStr, grpcplugin.TypeStr},
 			pipeline: configmodels.Pipelines{
 				"traces": {
 					InputType: configmodels.TracesDataType,
 					Receivers: []string{"jaeger"},
-					Exporters: []string{cassandra.TypeStr, elasticsearch.TypeStr, grpcExp.TypeStr},
+					Exporters: []string{cassandra.TypeStr, elasticsearch.TypeStr, grpcplugin.TypeStr},
 				},
 			},
 		},
@@ -155,6 +155,7 @@ func TestDefaultCollectorConfig(t *testing.T) {
 			}
 			sort.Strings(types)
 			assert.Equal(t, test.exporterTypes, types)
+			sort.Strings(cfg.Service.Pipelines["traces"].Receivers)
 			assert.EqualValues(t, test.pipeline, cfg.Service.Pipelines)
 		})
 	}
@@ -246,7 +247,7 @@ func TestDefaultIngesterConfig(t *testing.T) {
 					"traces": &configmodels.Pipeline{
 						InputType: configmodels.TracesDataType,
 						Receivers: []string{kafkaRec.TypeStr},
-						Exporters: []string{cassandra.TypeStr, elasticsearch.TypeStr, grpcExp.TypeStr},
+						Exporters: []string{cassandra.TypeStr, elasticsearch.TypeStr, grpcplugin.TypeStr},
 					},
 				},
 			},
